payment move to type guard
